renamed MessageProcessorActivator to RequestReplyProducer
